handle linked requests to FindAll of related struct
routes for loading linked resources are now automatically generated.
When calling a linked route, the request is automatically directed at
the FindAll method of the referenced resource with a query Parameter
that contains the reference ID with that the results must be filtered.

for example:
GET /posts/1/comments

redirects to CommentsResource FindAll method with Query Parameter
PostsID 1