Fix Unmarshalling of sql Null-Types #109
to marshal/unmarshal the sql Null-Types it is recommended to use the
library: gopkg.in/guregu/null.v2/zero

Otherwise it would be necessary to implement the required interfaces.

Unmarshalling is a litle bid dirty now because of the heavy reflection
usage. The UnmarshalJSON Methods of the zero package are currently not
used and a custom type switch is made.