new feature to rename a struct
the current situation always generated the structs name from it's type. So
a UnicornPost struct would have the generated name `unicornPosts`. But maybe,
a user wants to use a different naming scheme, like `unicorn-posts`. This is
now possible.

Maybe we should change the Jsonify and Dejsonify methods in the future that
they automatically generate *hyphenized* names of everything.