It's now possible to skip the `data` field for relations
This is useful if a user only wants to include the `links` without `data`.
Sometimes loading relationships is a very expensive operation. If relations
are not needed anyway, it is enough to send the links. A client can then
access the relations at a later time via the provided links.